Wait for fresh state in WaitForPlanExecution
If the controller status and robot state messages are received simultaneously
after a gap in comms, it was possible to start planning from a stale robot
state. Now we wait for a state message with a utime after the utime of the
controller status message that indicated the end of plan execution.