remove pelvisLink arg from IKServer.runIkTraj
I don't like that planning specific args were added to runIkTraj, we
should try to remove these.  As a start, I've moved the pelvis link
arg to the constructor of IKServer.  In the long run, the IKServer
should be much more generic and all the specific code needs to be in
class specific planners that only use IKServer as an interface to
matlab.