import iris earlier to avoid import error
On some systems we are getting the import error:

ImportError: dlopen: cannot load any more object with static TLS

which seems to be coming from the swig layer of iris.  The issue is
avoided by importing iris earlier.  It seems to be interacting in a
strange way with other modules loaded by director.