Create utility class to aggregate lidar data
Create ddLidarPointCloudLCM which listens to lidar messages and
aggregates the data. This provides a mechanism for users to obtain a
complete "snapshot" of a sensor's coverage, with the most recent data
for the area of coverage.

This isn't quite complete yet; proper aggregation for bidirectional
sensors is not yet implemented, nor is extraction of non-position data
to vtkPolyData.