Include channel name in message callbacks
Modify Python LCM callback logic so that, by default, callbacks are
called with the name of the channel on which the message was received
(falling back to the single argument callback in case of error, for
backwards compatibility). This is important for callbacks registered on
wildcard subscriptions (as will be needed for handling lidar).