remove usage of child frame objects for links in drake-visualizer
The callbacks associated with child frames do not scale well for greater
than a few hundred instances, and there are now users of drake visualizer
that are generating hundreds of links.

Perhaps in the future I'll optimize the frame callbacks, but for now I am
just removing child frames from the implementation.