support multiple geometries in the same treeviewer item, and add tests
This involved reorganizing the treeviewer internals a bit, but the
change is for the better. Now we attach the color and channel properties
directly to the polyData when it's created from the message data, rather
than waiting to attach it to the geometry item. This also means that
Geometry.createGeometry now always returns exactly one Geometry rather
than a list of some length (usually 1). With this, we can easily handle
the "geometries" field in the treeviewer protocol